requirements.txt
gspread==6.1.4
python-dotenv==1.0.1
pandas==2.2.3
google-generativeai==0.8.3